+++ packages/client/components/ActivityLibrary/ActivityDetailsSidebar.tsx
@@ -89,9 +89,6 @@ const ActivityDetailsSidebar = (props: Props) => {
         orgId
         organization {
           name
-          featureFlags {
-            recurringRetros
-          }
         }
         retroSettings: meetingSettings(meetingType: retrospective) {
           ...NewMeetingSettingsToggleCheckIn_settings
@@ -122,9 +119,9 @@ const ActivityDetailsSidebar = (props: Props) => {
     selectedTemplate.scope === 'PUBLIC'
       ? teams
       : selectedTemplate.scope === 'ORGANIZATION'
-      ? teams.filter((team) => team.orgId === selectedTemplate.orgId)
-      : // it is a team-scoped template, templateTeam  must exist
-        [templateTeam!]
+        ? teams.filter((team) => team.orgId === selectedTemplate.orgId)
+        : // it is a team-scoped template, templateTeam  must exist
+          [templateTeam!]
 
   const availableTeamsRef = useRef(availableTeams)
 
@@ -246,15 +243,13 @@ const ActivityDetailsSidebar = (props: Props) => {
     type === 'retrospective'
       ? 'Retro'
       : type === 'teamPrompt'
-      ? 'Standup'
-      : type === 'poker'
-      ? 'Poker'
-      : type === 'action'
-      ? 'Check-in'
-      : 'Meeting'
-  const withRecurrence =
-    type === 'teamPrompt' ||
-    (selectedTeam.organization.featureFlags.recurringRetros && type === 'retrospective')
+        ? 'Standup'
+        : type === 'poker'
+          ? 'Poker'
+          : type === 'action'
+            ? 'Check-in'
+            : 'Meeting'
+  const withRecurrence = type === 'teamPrompt' || type === 'retrospective'
 
   return (
     <>