+++ packages/client/components/Tag/TierTag.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled'
 import React from 'react'
 import {PALETTE} from '../../styles/paletteV3'
 import {TierLabel} from '../../types/constEnums'
-import {TierEnum} from '../../__generated__/NewMeetingQuery.graphql'
+import {TierEnum} from '../../__generated__/OrganizationSubscription.graphql'
 import BaseTag from './BaseTag'
 
 interface Props {