+++ packages/server/graphql/public/types/OrganizationFeatureFlags.ts
@@ -14,7 +14,7 @@ const OrganizationFeatureFlags: OrganizationFeatureFlagsResolvers = {
   kudos: ({kudos}) => !!kudos,
   aiIcebreakers: ({aiIcebreakers}) => !!aiIcebreakers,
   aiTemplate: ({aiTemplate}) => !!aiTemplate,
-  recurringRetros: ({recurringRetros}) => !!recurringRetros
+  relatedDiscussions: ({relatedDiscussions}) => !!relatedDiscussions
 }
 
 export default OrganizationFeatureFlags